+++ howhowSpecialProject/src/main/java/com/howhow/ecpayment/ECPaymentResult.java
@@ -1,0 +1,5 @@
+package com.howhow.ecpayment;
+
+public class ECPaymentResult {
+
+}
